# web/project-site/public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Audio transcription service using WhisperX",
        "title": "Quill API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/notes/{note_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a note by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "note_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Note"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the content of a note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Update a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "note_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Note update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.NoteUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Note"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a note by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Delete a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "note_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/transcription/{id}/notes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all notes attached to a transcription, ordered by time and creation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "List notes for a transcription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transcription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Note"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new note attached to the specified transcription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Create a note for a transcription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transcription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Note create payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.NoteCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Note"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.NoteCreateRequest": {
            "type": "object",
            "required": [
                "content",
                "quote"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 1
                },
                "end_time": {
                    "type": "number",
                    "minimum": 0
                },
                "end_word_index": {
                    "type": "integer",
                    "minimum": 0
                },
                "quote": {
                    "type": "string",
                    "minLength": 1
                },
                "start_time": {
                    "type": "number",
                    "minimum": 0
                },
                "start_word_index": {
                    "description": "Use gte=0 so 0 is valid (first word/time); avoid 'required' which fails for zero values",
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "api.NoteUpdateRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "uploaded",
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusUploaded",
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
        "models.MultiTrackFile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "description": "Original filename (used as speaker name)",
                    "type": "string"
                },
                "file_path": {
                    "description": "Full path to audio file",
                    "type": "string"
                },
                "gain": {
                    "description": "Gain value from .aup file",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "mute": {
                    "description": "Whether track is muted",
                    "type": "boolean"
                },
                "offset": {
                    "description": "Offset in seconds from .aup file",
                    "type": "number"
                },
                "pan": {
                    "description": "Pan value from .aup file (-1.0 to 1.0)",
                    "type": "number"
                },
                "track_index": {
                    "description": "Order of the track",
                    "type": "integer"
                },
                "transcription_job": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TranscriptionJob"
                        }
                    ]
                },
                "transcription_job_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Note": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "The user's note content (markdown/plain)",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_time": {
                    "type": "number"
                },
                "end_word_index": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "quote": {
                    "description": "The exact quoted text chosen by the user",
                    "type": "string"
                },
                "start_time": {
                    "description": "Time bounds for the selection (in seconds)",
                    "type": "number"
                },
                "start_word_index": {
                    "description": "Indexed selection into transcript by word positions",
                    "type": "integer"
                },
                "transcription": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TranscriptionJob"
                        }
                    ]
                },
                "transcription_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.TranscriptionJob": {
            "type": "object",
            "properties": {
                "audio_path": {
                    "type": "string"
                },
                "aup_file_path": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "diarization": {
                    "type": "boolean"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "individual_transcripts": {
                    "description": "JSON-serialized map[string]*string",
                    "type": "string"
                },
                "is_multi_track": {
                    "type": "boolean"
                },
                "merge_error": {
                    "type": "string"
                },
                "merge_status": {
                    "description": "none, pending, processing, completed, failed",
                    "type": "string"
                },
                "merged_audio_path": {
                    "type": "string"
                },
                "multi_track_files": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MultiTrackFile"
                    }
                },
                "multi_track_folder": {
                    "type": "string"
                },
                "parameters": {
                    "description": "WhisperX parameters",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.WhisperXParams"
                        }
                    ]
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transcript": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.WhisperXParams": {
            "type": "object",
            "properties": {
                "align_model": {
                    "description": "Alignment settings",
                    "type": "string"
                },
                "api_key": {
                    "description": "OpenAI settings",
                    "type": "string"
                },
                "attention_context_left": {
                    "description": "NVIDIA Parakeet-specific parameters for long-form audio",
                    "type": "integer"
                },
                "attention_context_right": {
                    "type": "integer"
                },
                "batch_size": {
                    "type": "integer"
                },
                "beam_size": {
                    "type": "integer"
                },
                "best_of": {
                    "type": "integer"
                },
                "callback_url": {
                    "description": "Webhook settings",
                    "type": "string"
                },
                "chunk_size": {
                    "type": "integer"
                },
                "compression_ratio_threshold": {
                    "type": "number"
                },
                "compute_type": {
                    "type": "string"
                },
                "condition_on_previous_text": {
                    "type": "boolean"
                },
                "device": {
                    "description": "Device and computation",
                    "type": "string"
                },
                "device_index": {
                    "type": "integer"
                },
                "diarize": {
                    "description": "Diarization settings",
                    "type": "boolean"
                },
                "diarize_model": {
                    "description": "Options: 'pyannote', 'nvidia_sortformer'",
                    "type": "string"
                },
                "fp16": {
                    "type": "boolean"
                },
                "hf_token": {
                    "description": "Token and progress",
                    "type": "string"
                },
                "highlight_words": {
                    "type": "boolean"
                },
                "initial_prompt": {
                    "type": "string"
                },
                "interpolate_method": {
                    "type": "string"
                },
                "is_multi_track_enabled": {
                    "description": "Multi-track transcription settings",
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "length_penalty": {
                    "type": "number"
                },
                "logprob_threshold": {
                    "type": "number"
                },
                "max_line_count": {
                    "type": "integer"
                },
                "max_line_width": {
                    "description": "Output formatting",
                    "type": "integer"
                },
                "max_new_tokens": {
                    "description": "Voxtral settings",
                    "type": "integer"
                },
                "max_speakers": {
                    "type": "integer"
                },
                "min_speakers": {
                    "type": "integer"
                },
                "model": {
                    "description": "Model parameters",
                    "type": "string"
                },
                "model_cache_only": {
                    "type": "boolean"
                },
                "model_dir": {
                    "type": "string"
                },
                "model_family": {
                    "description": "Model family (whisper or nvidia)",
                    "type": "string"
                },
                "no_align": {
                    "type": "boolean"
                },
                "no_speech_threshold": {
                    "type": "number"
                },
                "output_format": {
                    "description": "Output settings",
                    "type": "string"
                },
                "patience": {
                    "type": "number"
                },
                "print_progress": {
                    "type": "boolean"
                },
                "return_char_alignments": {
                    "type": "boolean"
                },
                "segment_resolution": {
                    "type": "string"
                },
                "speaker_embeddings": {
                    "type": "boolean"
                },
                "suppress_numerals": {
                    "type": "boolean"
                },
                "suppress_tokens": {
                    "type": "string"
                },
                "task": {
                    "description": "Task and language",
                    "type": "string"
                },
                "temperature": {
                    "description": "Transcription quality settings",
                    "type": "number"
                },
                "temperature_increment_on_fallback": {
                    "type": "number"
                },
                "threads": {
                    "type": "integer"
                },
                "vad_method": {
                    "description": "VAD (Voice Activity Detection) settings",
                    "type": "string"
                },
                "vad_offset": {
                    "type": "number"
                },
                "vad_onset": {
                    "type": "number"
                },
                "verbose": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token with Bearer prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}